Rename LockFactory to Session and move to its own submodule
